+++ firestore_api_flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.2
+
+* Support new metadata parameter access (supports fire_api 1.1.5+)
+* Supports docChanges added,removed,modified in collection streams
+
 ## 1.1.1
 
 * Fix transactions